+++ src/graph_pes/graph_pes_model.py
@@ -97,7 +97,7 @@ class GraphPESModel(nn.Module, ABC):
 
     For more details on how these are calculated, see :doc:`../theory`.
 
-    :class:`~graph_pes.GraphPESModel` objects save various peices of extra
+    :class:`~graph_pes.GraphPESModel` objects save various pieces of extra
     metadata to the ``state_dict`` via the
     :meth:`~graph_pes.GraphPESModel.get_extra_state` and
     :meth:`~graph_pes.GraphPESModel.set_extra_state` methods.
@@ -230,7 +230,7 @@ class GraphPESModel(nn.Module, ABC):
             # both the cell and the atomic positions).
             #
             # F. Knuth et al. All-electron formalism for total energy strain
-            # derivatives and stress tensor components for numeric atom-centered
+            # derivatives and stress tensor components for numeric atom-centred
             # orbitals. Computer Physics Communications 190, 33–50 (2015).
 
             change_to_cell = torch.zeros_like(graph.cell)
@@ -370,7 +370,7 @@ class GraphPESModel(nn.Module, ABC):
         This method does two things:
 
         1. iterates over all the model's :class:`~torch.nn.Module` components
-           (inlcuding itself) and calls their :meth:`pre_fit` method (if it exists -
+           (including itself) and calls their :meth:`pre_fit` method (if it exists -
            see for instance :class:`~graph_pes.models.LearnableOffset` for
            an example of a model-specific pre-fit method, and
            :class:`~graph_pes.models.components.scaling.LocalEnergiesScaler` for
@@ -575,7 +575,10 @@ class GraphPESModel(nn.Module, ABC):
 
     @torch.jit.unused
     def ase_calculator(
-        self, device: torch.device | str | None = None, skin: float = 1.0
+        self,
+        device: torch.device | str | None = None,
+        skin: float = 1.0,
+        cache_threebody: bool = True,
     ) -> "GraphPESCalculator":
         """
         Return an ASE calculator wrapping this model. See
@@ -592,10 +595,19 @@ class GraphPESModel(nn.Module, ABC):
             than half of this distance between calls to `calculate`, the
             neighbour list will be reused, saving (in some cases) significant
             computation time.
+        cache_threebody
+            Whether to cache the three-body neighbour list entries. In many
+            cases, this can accelerate MD simulations by avoiding these quite
+            expensive recalculations. Tuning the ``skin`` parameter is important
+            to optimise the trade-off between less frequent but more expensive
+            neighbour list recalculations. This options is ignored
+            if the model does not use three-body interactions.
         """
         from graph_pes.utils.calculator import GraphPESCalculator
 
-        return GraphPESCalculator(self, device=device, skin=skin)
+        return GraphPESCalculator(
+            self, device=device, skin=skin, cache_threebody=cache_threebody
+        )
 
     @torch.jit.unused
     def torch_sim_model(