+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Implement Orb model
+
 ## [0.2.2] - 2025-10-07
 
 Fixed bugs in the MACE-OMOL interface.